All the file add

+++ Personal_Diary/diary_app/views.py
@@ -7,19 +7,33 @@ from .models import Category, Notebook, Expense, RegisterUser
 from django.contrib.auth.hashers import check_password, make_password
 from django.contrib.auth.models import AbstractUser
 from django.db import models
+from django.db.models import Sum
+from django.utils.dateparse import parse_date
+
+
 
 
 
 User = get_user_model()
+
+def about_page(request):
+    return render(request, 'diary_app/about.html')  
 
 def registration(request):
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
         email = request.POST.get('email')
+        re_enter_password= request.POST.get('password')
         hashed_password = make_password(password)
-        user = RegisterUser(username=username, password=hashed_password, email=email)
-        user.save()
+        hashed_password = make_password(re_enter_password)
+        if password==re_enter_password:
+
+            user = RegisterUser(username=username, password=hashed_password, email=email)
+            user.save()
+        else:
+            re_enter_password= request.POST.get('password')
+
         return redirect('login_view') 
     return render(request, 'diary_app/register.html')
 
@@ -59,25 +73,37 @@ def index(request):
 def add_category(request):
     if request.method == 'POST':
         category_name = request.POST.get('category_name')
-        Category.objects.create(user=request.user, category_name=category_name)
-        return redirect('index')
-    return render(request, 'diary_app/add_category.html')
+        if category_name:
+            Category.objects.create(category_name=category_name, user=request.user)
+            return redirect('add_category')  # or wherever you want to redirect
+
+    categories = Category.objects.filter(user=request.user).order_by('category_name')
+    return render(request, 'diary_app/add_category.html', {'categories': categories})
 
 @login_required
 def edit_category(request, category_id):
-    category = Category.objects.get(id=category_id, user=request.user)
+    category = get_object_or_404(Category, id=category_id, user=request.user)
+
     if request.method == 'POST':
         category_name = request.POST.get('category_name')
-        category.category_name = category_name
-        category.save()
-        return redirect('index')
+        if category_name:
+            category.category_name = category_name
+            category.save()
+            return redirect('add_category')
+
     return render(request, 'diary_app/edit_category.html', {'category': category})
+
 
 @login_required
 def delete_category(request, category_id):
-    category = Category.objects.get(id=category_id, user=request.user)
-    category.delete()
-    return redirect('index')
+    category = get_object_or_404(Category, id=category_id, user=request.user)
+
+    if request.method == 'POST':
+        category.delete()
+        return redirect('add_category')
+
+    return render(request, 'diary_app/delete_category.html', {'category': category})
+
 
 @login_required
 def add_notebook(request):
@@ -153,27 +179,38 @@ def expenses_by_category(request, category_id):
     })
 
 @login_required
-def notebook_detail(request, notebook_id):
-    notebook = Notebook.objects.get(id=notebook_id, user=request.user)
-    return render(request, 'diary_app/notebook_detail.html', {'notebook': notebook})
+def my_notebooks(request):
+    notebooks = Notebook.objects.filter(user=request.user).order_by('-created_at')
+    return render(request, 'diary_app/my_notebooks.html', {'notebooks': notebooks})
 
-@login_required
-def expense_detail(request, expense_id):
-    expense = Expense.objects.get(id=expense_id, user=request.user)
-    return render(request, 'diary_app/expense_detail.html', {'expense': expense})
 
 @login_required
 def my_expenses_view(request):
     categories = Category.objects.filter(user=request.user)
     selected_category = request.GET.get('category')
+    from_date = request.GET.get('from_date')
+    to_date = request.GET.get('to_date')
+
+    expenses = Expense.objects.filter(user=request.user)
 
     if selected_category:
-        expenses = Expense.objects.filter(user=request.user, category__id=selected_category).order_by('-created_at')
-    else:
-        expenses = Expense.objects.filter(user=request.user).order_by('-created_at')
+        expenses = expenses.filter(category__id=selected_category)
+
+    if from_date:
+        expenses = expenses.filter(created_at__date__gte=from_date)
+    if to_date:
+        expenses = expenses.filter(created_at__date__lte=to_date)
+
+    expenses = expenses.order_by('-created_at')
+    total_amount = expenses.aggregate(total=Sum('amount'))['total'] or 0
 
     return render(request, 'diary_app/my_expenses.html', {
         'expenses': expenses,
         'categories': categories,
-        'selected_category': selected_category
+        'selected_category': selected_category,
+        'total_amount': total_amount,
+        'from_date': from_date,
+        'to_date': to_date
     })
+
+
